// src/utils/types.ts
export type BookType = {
  title: string;
  price: string;
  image: string;
  url: string;
  UPC: string;
  "Product Type": string;
  "Price (excl. tax)": string;
  "Price (incl. tax)": string;
  Tax: string;
  Availability: string;
  "Number of reviews": string;
  rating: string;
  category: string;
  description: string | null;
};